+++ Sources/SwiftUIKit/Extensions/Cocoa/NSImage+ImageResource.swift
@@ -1,0 +1,17 @@
+//
+//  NSImage+ImageResource.swift
+//  SwiftUIKit
+//
+//  Created by Daniel Saidi on 2021-07-25.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+#if os(macOS)
+import Cocoa
+
+/**
+ This typealias makes it easy to use NSImage and other image
+ types interchangeably, where `ImageResource` is supported.
+ */
+public typealias ImageResource = NSImage
+#endif