+++ Sources/SwiftUIKit/Extensions/UIKit/UIImage+ImageResource.swift
@@ -1,0 +1,17 @@
+//
+//  UIImage+ImageResource.swift
+//  SwiftUIKit
+//
+//  Created by Daniel Saidi on 2021-07-25.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+#if os(iOS)
+import UIKit
+
+/**
+ This typealias makes it easy to use UIImage and other image
+ types interchangeably, where `ImageResource` is supported.
+ */
+public typealias ImageResource = UIImage
+#endif